Fixed the whole thing with rejoining the patches
multiplied the output array with 255.0 to renormalize and now the image isn't 'black' anymore. And also trashed my custom rejoining code with that of sklearn, which makes a huge no. of patches with stride 1. So now the output image is hugely more clear than was before. But even now, the result is not very admirable. Just that the attempt fell good now that it 'works'.